Added 2 new switches:  mAspectCorrect and mAspectNoStretch
if mAspectCorrect is false, then the game will change desktop resolution to that of the game and no aspect correction will be done.

if mAspectNoStretch is true and mAspectCorrect is true, then the game will attempt to find a widescreen resolution whose height matches the requested height.

Both values are written to the registry.  If mAspectCorrect is written to the registry, then the registry value will override the value you give it in the application Ctor.

mAspectNoStretch is written to the registry, but for testing purposes, this value is not read from registry.  The application Ctor value will override the registry.  Uncomment line 1922 in SexyAppBase.cpp to change this behavior.

Also added more functions to BassLoader to get track time information (in seconds).

git-svn-id: https://subversion.assembla.com/svn/sexy/trunk@17 037ef64b-6aa3-42ed-9b66-4ad3d2b4a348